Add all topic names to container props and re combine default mode tests
